// src/types/common.d.ts
export type LostPet = {
  callName: string;
  callTel: string;
  happenDt: string;
  happenAddr: string;
  happenAddrDtl?: string;
  happenPlace: string;
  orgNm: string;
  popfile: string;
  kindCd: string;
  colorCd: string;
  sexCd: string;
  age: string;
  specialMark: string;
  rfidCd: string;
};